// program/clients/typescript/src/generated/types/status.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from '@solana/kit';

export enum Status {
  Paid,
  Cleared,
  Refunded,
}

export type StatusArgs = Status;

export function getStatusEncoder(): FixedSizeEncoder<StatusArgs> {
  return getEnumEncoder(Status);
}

export function getStatusDecoder(): FixedSizeDecoder<Status> {
  return getEnumDecoder(Status);
}

export function getStatusCodec(): FixedSizeCodec<StatusArgs, Status> {
  return combineCodec(getStatusEncoder(), getStatusDecoder());
}